# docs/reference/openapi-spec.yaml
swagger: "2.0"
info:
  title: "CLS Backend API"
  description: "Cluster Lifecycle Service Backend - Simplified Single-Tenant Architecture"
  version: "1.0.0"
  contact:
    name: "CLS Backend Team"
    url: "https://github.com/your-org/cls-backend"
  license:
    name: "MIT"
    url: "https://opensource.org/licenses/MIT"
host: "YOUR_API_GATEWAY_HOST"
basePath: "/api/v1"
schemes:
  - "https"
consumes:
  - "application/json"
produces:
  - "application/json"

# Security definitions for Google Cloud user authentication
securityDefinitions:
  # Google OAuth2 with user authentication
  google_oauth2:
    type: oauth2
    authorizationUrl: https://accounts.google.com/o/oauth2/auth
    flow: implicit
    x-google-jwks_uri: https://www.googleapis.com/oauth2/v3/certs
    x-google-audiences: "32555940559.apps.googleusercontent.com"
    scopes:
      openid: OpenID Connect scope
      email: Access to user email address
      profile: Access to user profile information
    description: |
      Google OAuth2 authentication with email scope.
      Users must be authenticated via gcloud and have valid access tokens.

  # For development/testing (optional fallback)
  api_key:
    type: "apiKey"
    name: "key"
    in: "query"

# Global security requirement - Google OAuth2 required for production
security:
  - google_oauth2: ["openid", "email", "profile"]

# Path definitions
paths:
...
  /clusters:
    get:
      summary: "List Clusters"
      description: "List all clusters with optional filtering and pagination"
      operationId: "listClusters"
      security:
        - google_oauth2: ["openid", "email", "profile"]
      parameters:
        - name: "limit"
          in: "query"
          type: "integer"
          minimum: 1
          maximum: 100
          default: 50
          description: "Maximum number of results"
        - name: "offset"
          in: "query"
          type: "integer"
          minimum: 0
          default: 0
          description: "Number of results to skip"
        - name: "platform"
          in: "query"
          type: "string"
          enum: ["gcp", "aws", "azure"]
          description: "Filter by platform type"
        - name: "status"
          in: "query"
          type: "string"
          enum: ["Pending", "Progressing", "Ready", "Failed"]
          description: "Filter by status phase"
        - name: "X-User-Email"
          in: "header"
          type: "string"
          required: true
          description: "User email for authorization"
      responses:
        200:
          description: "List of clusters"
          schema:
            $ref: "#/definitions/ClusterListResponse"
        400:
          description: "Invalid request parameters"
          schema:
            $ref: "#/definitions/ErrorResponse"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/ErrorResponse"
      x-google-backend:
        address: "YOUR_BACKEND_SERVICE_URL"
        path_translation: "APPEND_PATH_TO_ADDRESS"

    post:
      summary: "Create Cluster"
      description: "Create a new cluster"
      operationId: "createCluster"
      security:
        - google_oauth2: ["openid", "email", "profile"]
      parameters:
        - name: "X-User-Email"
          in: "header"
          type: "string"
          required: true
          description: "User email for authorization"
        - name: "body"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/CreateClusterRequest"
      responses:
        201:
          description: "Cluster created successfully"
          schema:
            $ref: "#/definitions/Cluster"
        400:
          description: "Invalid request body"
          schema:
            $ref: "#/definitions/ErrorResponse"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/ErrorResponse"
        409:
          description: "Cluster already exists"
          schema:
            $ref: "#/definitions/ErrorResponse"
      x-google-backend:
        address: "YOUR_BACKEND_SERVICE_URL"
        path_translation: "APPEND_PATH_TO_ADDRESS"
...
definitions:
...
  ErrorResponse:
    type: "object"
    required:
      - "error"
    properties:
      error:
        type: "object"
        required:
          - "code"
          - "message"
        properties:
          code:
            type: "string"
          message:
            type: "string"
          details:
            type: "string"

  # Cluster definitions
  Cluster:
    type: "object"
    required:
      - "id"
      - "name"
      - "generation"
      - "resource_version"
      - "spec"
      - "status"
      - "created_at"
      - "updated_at"
      - "created_by"
    properties:
      id:
        type: "string"
        format: "uuid"
      name:
        type: "string"
        minLength: 1
        maxLength: 63
        pattern: "^[a-z0-9-]+$"
      generation:
        type: "integer"
        format: "int64"
        minimum: 1
      resource_version:
        type: "string"
      spec:
        $ref: "#/definitions/ClusterSpec"
      status:
        $ref: "#/definitions/ClusterStatus"
      created_at:
        type: "string"
        format: "date-time"
      updated_at:
        type: "string"
        format: "date-time"
      created_by:
        type: "string"
        format: "email"
  ClusterSpec:
    type: "object"
    required:
      - "platform"
    properties:
      infraID:
        type: "string"
      platform:
        $ref: "#/definitions/PlatformSpec"
      release:
        $ref: "#/definitions/ReleaseSpec"
      networking:
        $ref: "#/definitions/NetworkingSpec"
      dns:
        $ref: "#/definitions/DNSSpec"
  PlatformSpec:
    type: "object"
    required:
      - "type"
    properties:
      type:
        type: "string"
        enum: ["gcp", "aws", "azure"]
      gcp:
        $ref: "#/definitions/GCPPlatformSpec"
      aws:
        $ref: "#/definitions/AWSPlatformSpec"
      azure:
        $ref: "#/definitions/AzurePlatformSpec"
  GCPPlatformSpec:
    type: "object"
    properties:
      projectID:
        type: "string"
      region:
        type: "string"
      zone:
        type: "string"
      network:
        type: "string"
      subnet:
        type: "string"
  AWSPlatformSpec:
    type: "object"
    properties:
      region:
        type: "string"
      availabilityZones:
        type: "array"
        items:
          type: "string"
      vpcId:
        type: "string"
      subnetIds:
        type: "array"
        items:
          type: "string"
  AzurePlatformSpec:
    type: "object"
    properties:
      region:
        type: "string"
      resourceGroup:
        type: "string"
      vnetName:
        type: "string"
      subnetName:
        type: "string"
  ReleaseSpec:
    type: "object"
    properties:
      image:
        type: "string"
      version:
        type: "string"
  NetworkingSpec:
    type: "object"
    properties:
      clusterNetwork:
        type: "array"
        items:
          $ref: "#/definitions/ClusterNetworkEntry"
      serviceNetwork:
        type: "array"
        items:
          type: "string"
  ClusterNetworkEntry:
    type: "object"
    properties:
      cidr:
        type: "string"
      hostPrefix:
        type: "integer"
  DNSSpec:
    type: "object"
    properties:
      baseDomain:
        type: "string"
  ClusterStatus:
    type: "object"
    required:
      - "observedGeneration"
      - "conditions"
      - "phase"
      - "lastUpdateTime"
    properties:
      observedGeneration:
        type: "integer"
        format: "int64"
      conditions:
        type: "array"
        items:
          $ref: "#/definitions/Condition"
      phase:
        type: "string"
        enum: ["Pending", "Progressing", "Ready", "Failed"]
      message:
        type: "string"
      reason:
        type: "string"
      lastUpdateTime:
        type: "string"
        format: "date-time"
  Condition:
    type: "object"
    required:
      - "type"
      - "status"
      - "lastTransitionTime"
    properties:
      type:
        type: "string"
      status:
        type: "string"
        enum: ["True", "False", "Unknown"]
      lastTransitionTime:
        type: "string"
        format: "date-time"
      reason:
        type: "string"
      message:
        type: "string"
  CreateClusterRequest:
    type: "object"
    required:
      - "name"
      - "spec"
    properties:
      name:
        type: "string"
        minLength: 1
        maxLength: 63
        pattern: "^[a-z0-9-]+$"
      spec:
        $ref: "#/definitions/ClusterSpec"
      target_project_id:
        type: "string"
...
  ClusterListResponse:
    type: "object"
    required:
      - "clusters"
      - "total"
      - "limit"
      - "offset"
    properties:
      clusters:
        type: "array"
        items:
          $ref: "#/definitions/Cluster"
      total:
        type: "integer"
      limit:
        type: "integer"
      offset:
        type: "integer"
...